+++ Cargo.toml
@@ -3,9 +3,13 @@ name = "RustyBear-Engine"
 version = "0.1.0"
 edition = "2021"
 
+[lib]
+crate-type = ["cdylib", "rlib"]
+
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+cfg-if = "1"
 bimap = { version = "0.6.3", features = ["serde"] }
 env_logger = "0.10.0"
 gilrs = "0.10.2"
@@ -18,3 +22,17 @@ simplelog = "0.12.1"
 wgpu = "0.16.1"
 winit = "0.28.6"
 winit-fullscreen = "1.0.1"
+instant = "0.1"
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+console_error_panic_hook = "0.1.6"
+console_log = "1.0"
+wgpu = { version = "0.16", features = ["webgl"]}
+wasm-bindgen = "0.2"
+wasm-bindgen-futures = "0.4.30"
+web-sys = { version = "0.3", features = [
+    "Document",
+    "Window",
+    "Element",
+]}
+instant = { version = "0.1", features = [ "wasm-bindgen" ] }
