+++ index.html
@@ -1,0 +1,22 @@
+<!DOCTYPE html>
+
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="Viewport" content="width=device-width, initial-scale=1.0">
+    <title>RustyBear-Engine Sandbox</title>
+    <style>
+        canvas {
+            background-color: black;
+        }
+    </style>
+</head>
+
+<body id="rusty-bear-engine">
+    <script type="module">
+        import init from "./pkg/RustyBear_Engine.js";
+        init().then(() => {
+            console.log("Loaded RustyBear-Engine.");
+        });
+    </script>
+</body>
